+++ sync_telegram/models/mail/__init__.py
@@ -1,0 +1,3 @@
+# License MIT (https://opensource.org/licenses/MIT).
+from . import mail_channel
+from . import res_users_settings