+++ sync/doc/changelog.rst
@@ -1,3 +1,16 @@
+`6.1.0`
+-------
+- **Improvement:** use light telegram library in Telegram demo project (pyTelegramBotAPI)
+- **Fix:** add button to show secret parameter value
+- **Improvement:** add eval context function: `url2bin`
+- **Fix:** Project export tool: add records of `sync.project.context`
+- **Fix:** properly cleanup `base.automation` and action records
+
+`6.0.1`
+-------
+- **Fix:** error on empty eval context in a project
+
+
 `6.0.0`
 -------